+++ src/main/java/ironfurnaces/loaders/PacketInit.java
@@ -11,6 +11,7 @@ public class PacketInit {
         NetworkUtils.registerPacket(S2CSyncInstancesToMenuPackets::new);
         NetworkUtils.registerPacket(S2CSyncPatternAndStatsToMenuPackets::new);
         NetworkUtils.registerPacket(S2CSyncFDDataPacket::new);
+        NetworkUtils.registerPacket(S2CSyncAugmentPacket::new);