+++ src/main/java/ironfurnaces/tileentity/furnaces/menu/FurnacePatternMenu.java
@@ -248,7 +248,6 @@ public class FurnacePatternMenu extends DistributePartitionContainerMenu {
                 addSlotListener(containerListener);
             }
         }
-        System.out.println("menu init");
 
     }